+++ packages/core/src/model/chain/optimism.ts
@@ -5,11 +5,16 @@ const testnetExplorerUrl = 'https://kovan-optimistic.etherscan.io'
 
 export const OptimismKovan: Chain = {
   chainId: 69,
-  chainName: 'OptimismKovan',
+  chainName: 'Optimism Kovan',
   isTestChain: true,
   isLocalChain: false,
   multicallAddress: '0xE71bf4622578c7d1526A88CD3060f03030E99a04',
   rpcUrl: 'https://kovan.optimism.io',
+  nativeCurrency: {
+    name: 'Kovan Ether',
+    symbol: 'ETH',
+    decimals: 18,
+  },
   blockExplorerUrl: testnetExplorerUrl,
   getExplorerAddressLink: getAddressLink(testnetExplorerUrl),
   getExplorerTransactionLink: getTransactionLink(testnetExplorerUrl),
@@ -19,11 +24,16 @@ const testnetGoerliExplorerUrl = 'https://blockscout.com/optimism/goerli'
 
 export const OptimismGoerli: Chain = {
   chainId: 420,
-  chainName: 'OptimismGoerli',
+  chainName: 'Optimism Goerli',
   isTestChain: true,
   isLocalChain: false,
   multicallAddress: '0xC8315CC7DCDF57476a8a1D184505845d52711024',
   rpcUrl: 'https://goerli.optimism.io',
+  nativeCurrency: {
+    name: 'Goerli Ether',
+    symbol: 'ETH',
+    decimals: 18,
+  },
   blockExplorerUrl: testnetGoerliExplorerUrl,
   getExplorerAddressLink: getAddressLink(testnetGoerliExplorerUrl),
   getExplorerTransactionLink: getTransactionLink(testnetGoerliExplorerUrl),
@@ -38,6 +48,11 @@ export const Optimism: Chain = {
   isLocalChain: false,
   multicallAddress: '0x35A6Cdb2C9AD4a45112df4a04147EB07dFA01aB7',
   rpcUrl: 'https://mainnet.optimism.io',
+  nativeCurrency: {
+    name: 'Ether',
+    symbol: 'ETH',
+    decimals: 18,
+  },
   blockExplorerUrl: optimismExplorerUrl,
   getExplorerAddressLink: getAddressLink(optimismExplorerUrl),
   getExplorerTransactionLink: getTransactionLink(optimismExplorerUrl),
@@ -45,5 +60,6 @@ export const Optimism: Chain = {
 
 export default {
   OptimismKovan,
+  OptimismGoerli,
   Optimism,
 }