+++ packages/core/src/providers/chainState/common/chainStateReducer.ts
@@ -1,8 +1,5 @@
 import { ChainState } from './model'
 
-/**
- * @public
- */
 export interface State {
   [chainId: number]:
     | {