+++ packages/core/src/providers/network/connectors/implementations/coinbase.ts
@@ -1,4 +1,4 @@
-import { Connector, Update } from '../connector'
+import { Connector, ConnectorUpdateData } from '../connector'
 import { providers } from 'ethers'
 import detectEthereumProvider from '@metamask/detect-provider'
 import { Event } from '../../../../helpers/event'
@@ -28,8 +28,9 @@ export async function getCoinbaseProvider() {
 
 export class CoinbaseWalletConnector implements Connector {
   public provider?: providers.Web3Provider
+  public readonly name = 'CoinbaseWallet'
 
-  readonly update = new Event<Update>()
+  readonly update = new Event<ConnectorUpdateData>()
 
   private async init() {
     if (this.provider) return