+++ packages/core/src/providers/network/readonlyNetworks/model.ts
@@ -2,7 +2,6 @@ import { ChainId } from '../../../constants'
 import type { providers } from 'ethers'
 
 export interface NetworkState {
-  nonStaticCalls: number
   errors: Error[]
 }
 
@@ -20,16 +19,10 @@ export interface ReadonlyNetworksModel {
   networkStates: NetworkStates
 }
 
-export interface UpdateNonStaticCallsCount {
-  type: 'UPDATE_NON_STATIC_CALLS_COUNT'
-  chainId: ChainId
-  count: number
-}
-
 export interface PropagateChainError {
   type: 'ADD_ERROR'
   chainId: ChainId
   error: Error
 }
 
-export type Actions = UpdateNonStaticCallsCount | PropagateChainError
+export type Actions = PropagateChainError