+++ packages/core/src/providers/transactions/model.ts
@@ -1,8 +1,5 @@
 import type { TransactionReceipt, TransactionResponse } from '@ethersproject/abstract-provider'
 
-/**
- * @public
- */
 export interface StoredTransaction {
   transaction: TransactionResponse
   submittedAt: number
@@ -11,6 +8,8 @@ export interface StoredTransaction {
   transactionName?: string
   originalTransaction?: TransactionResponse
 }
+
+export type UpdatedTransaction = Omit<StoredTransaction, 'submittedAt'> & { receipt: TransactionReceipt }
 
 /**
  * @public
@@ -22,9 +21,6 @@ export function getStoredTransactionState(transaction: StoredTransaction) {
   return 'Mining'
 }
 
-/**
- * @public
- */
 export type StoredTransactions = {
   [chainId: number]: StoredTransaction[]
 }